Add tests and benchmark for byte array.
